+++ rollup.config.js
@@ -1,0 +1,40 @@
+import resolve from 'rollup-plugin-node-resolve';
+import commonjs from 'rollup-plugin-commonjs';
+import babel from 'rollup-plugin-babel';
+import pkg from './package.json';
+
+export default [
+	// browser-friendly UMD build
+	{
+		entry: 'src/js/main.iife.js',
+		dest: pkg.browser,
+		format: 'iife',
+		moduleName: 'chocolat',
+		plugins: [
+			resolve(), // so Rollup can find `jquery`
+			commonjs(), // so Rollup can convert `jquery` to an ES module
+			babel({
+				exclude: ['node_modules/**']
+			})
+		]
+	},
+
+	// CommonJS (for Node) and ES module (for bundlers) build.
+	// (We could have three entries in the configuration array
+	// instead of two, but it's quicker to generate multiple
+	// builds from a single configuration where possible, using
+	// the `targets` option which can specify `dest` and `format`)
+	{
+		entry: 'src/js/main.esm.js',
+		external: ['jquery', 'es6-promise/auto'],
+		targets: [
+			{ dest: pkg.main, format: 'cjs' },
+			{ dest: pkg.module, format: 'es' }
+		],
+		plugins: [
+			babel({
+				exclude: ['node_modules/**']
+			})
+		]
+	}
+];
